Removing the clear method from the layer
We'll want this back eventually.  But this lets us have a predictable structure
for the vector layer event object.  When we do add a clear method, we could add
a different event type that didn't come with features and extent.  At the same
time, we can give this vector event object a more specific name.